Update C examples: Use int now instead of int64_t for CALI_TYPE_INT values

+++ test/ci_app_tests/ci_test_c_snapshot.c
@@ -17,15 +17,12 @@ int main()
   cali_id_t event_val_attr =
     cali_create_attribute("int_arg",    CALI_TYPE_INT,    CALI_ATTR_ASVALUE);
 
-  int64_t     event_val = 42;
+  int         event_val = 42;
   const char* event_str = "teststring";
   
-  cali_id_t   event_attr[2] = { event_str_attr,
-                                event_val_attr  };
-  const void* event_data[2] = { event_str,
-                                &event_val      };
-  size_t      event_size[2] = { strlen(event_str),
-                                sizeof(int64_t) };
+  cali_id_t   event_attr[2] = { event_str_attr,    event_val_attr };
+  const void* event_data[2] = { event_str,         &event_val     };
+  size_t      event_size[2] = { strlen(event_str), sizeof(int)    };
 
   cali_push_snapshot(CALI_SCOPE_PROCESS | CALI_SCOPE_THREAD,
                      2, event_attr, event_data, event_size);
